Testing

+++ app/src/main/java/com/example/Calls/Dialog/QuestionDialog.java
@@ -10,12 +10,14 @@ import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatDialogFragment;
 import android.util.Log;
 
+import com.example.Calls.BackEnd.CutterFiles.FFmpegCutter;
 import com.example.Calls.BackEnd.Files.Directories;
 import com.example.Calls.BackEnd.Files.FileSystemParameters;
 import com.example.Calls.BackEnd.Mail.Mailer;
 import com.example.Calls.EditTextRecord;
 import com.example.Calls.MainActivity;
 import com.example.Calls.Model.Repositories.RecordRepository;
+import com.example.Calls.Play;
 import com.example.Calls.R;
 import com.example.Calls.SelectRecord;
 import com.example.Calls.WaitInEndPlay;
@@ -25,11 +27,23 @@ import java.io.File;
 @SuppressLint("ValidFragment")
 public class QuestionDialog extends AppCompatDialogFragment {
 
-    private EditTextRecord editTextRecord;
+    private EditTextRecord editTextRecord = null;
+    private WaitInEndPlay waitInEndPlay = null;
+    private Play play = null;
 
     @SuppressLint("ValidFragment")
     public QuestionDialog(EditTextRecord editTextRecord){
         this.editTextRecord = editTextRecord;
+    }
+
+    @SuppressLint("ValidFragment")
+    public QuestionDialog(WaitInEndPlay waitInEndPlay){
+        this.waitInEndPlay = waitInEndPlay;
+    }
+
+    @SuppressLint("ValidFragment")
+    public QuestionDialog(Play play){
+        this.play = play;
     }
 
     @Override
@@ -43,11 +57,22 @@ public class QuestionDialog extends AppCompatDialogFragment {
                 .setMessage("Вы уверены, что хотите закрыть полученный результат?")
                 .setPositiveButton("Да", new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int id) {
-                        Directories directories = new Directories(RecordRepository.getSelectedRecord());
-                        directories.deleteDirectory(new File(FileSystemParameters.getPathForSelectedRecord()));
+                        Intent intent = null;
+                        if(editTextRecord != null){
+                            intent = new Intent(editTextRecord, MainActivity.class);
+                            Directories directories = new Directories(RecordRepository.getSelectedRecord());
+                            directories.deleteDirectory(new File(FileSystemParameters.getPathForSelectedRecord()));
+                        }
+                        else if(waitInEndPlay != null){
+                            intent = new Intent(waitInEndPlay, MainActivity.class);
+                            FFmpegCutter.isStop = true;
+                        }
+                        else if(play != null){
+                            intent = new Intent(play, SelectRecord.class);
+                        }
 
-                        Intent intent = new Intent(editTextRecord, MainActivity.class);
                         startActivity(intent);
+
                         dialog.cancel();
                     }
                 })
